requirements.txt

tensorflow==2.12.0
numpy==1.24.3
absl-py==1.4.0
pillow==9.5.0
